+++ Robust.Shared/GameObjects/IComponentFactory.cs
@@ -51,7 +51,7 @@ namespace Robust.Shared.GameObjects
     /// <seealso cref="IComponent" />
     public interface IComponentFactory
     {
-        event Action<ComponentRegistration> ComponentAdded;
+        event Action<ComponentRegistration[]> ComponentsAdded;
         event Action<string> ComponentIgnoreAdded;
 
         /// <summary>
@@ -86,9 +86,8 @@ namespace Robust.Shared.GameObjects
         /// <summary>
         /// Registers a component name as being ignored.
         /// </summary>
-        /// <param name="name">The name to be ignored.</param>
-        /// <param name="overwrite">Whether to override existing settings instead of throwing an exception in the case of duplicates.</param>
-        void RegisterIgnore(string name, bool overwrite = false);
+        /// <param name="names">The names to be ignored.</param>
+        void RegisterIgnore(params string[] names);
 
         /// <summary>
         /// Disables throwing on missing components. Missing components will instead be treated as ignored.
@@ -270,6 +269,7 @@ namespace Robust.Shared.GameObjects
         /// </summary>
         void DoAutoRegistrations();
 
+        IEnumerable<ComponentRegistration> GetAllRegistrations();
         IEnumerable<CompIdx> GetAllRefTypes();
         void GenerateNetIds();