+++ Robust.UnitTesting/Shared/GameObjects/EntityEventBusTests.ComponentEvent.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using Moq;
 using NUnit.Framework;
@@ -80,7 +81,7 @@ namespace Robust.UnitTesting.Shared.GameObjects
             var compFacMock = new Mock<IComponentFactory>();
 
             compFacMock.Setup(m => m.GetRegistration(CompIdx.Index<MetaDataComponent>())).Returns(compRegistration);
-            compFacMock.Setup(m => m.GetAllRefTypes()).Returns(new[] { CompIdx.Index<MetaDataComponent>() });
+            compFacMock.Setup(m => m.GetAllRegistrations()).Returns(new[] { compRegistration });
             entManMock.Setup(m => m.ComponentFactory).Returns(compFacMock.Object);
 
             IComponent? outIComponent = compInstance;
@@ -136,7 +137,7 @@ namespace Robust.UnitTesting.Shared.GameObjects
             var compFacMock = new Mock<IComponentFactory>();
 
             compFacMock.Setup(m => m.GetRegistration(CompIdx.Index<MetaDataComponent>())).Returns(compRegistration);
-            compFacMock.Setup(m => m.GetAllRefTypes()).Returns(new[] { CompIdx.Index<MetaDataComponent>() });
+            compFacMock.Setup(m => m.GetAllRegistrations()).Returns(new[] { compRegistration });
             entManMock.Setup(m => m.ComponentFactory).Returns(compFacMock.Object);
 
             IComponent? outIComponent = compInstance;
@@ -181,7 +182,7 @@ namespace Robust.UnitTesting.Shared.GameObjects
             var entManMock = new Mock<IEntityManager>();
             var compFacMock = new Mock<IComponentFactory>();
 
-            List<CompIdx> allRefTypes = new();
+            List<ComponentRegistration> allRefTypes = new();
             void Setup<T>(out T instance) where T : IComponent, new()
             {
                 IComponent? inst = instance = new T();
@@ -194,14 +195,14 @@ namespace Robust.UnitTesting.Shared.GameObjects
                 entManMock.Setup(m => m.TryGetComponent(entUid, CompIdx.Index<T>(), out inst)).Returns(true);
                 entManMock.Setup(m => m.GetComponent(entUid, CompIdx.Index<T>())).Returns(inst);
                 entManMock.Setup(m => m.GetComponentInternal(entUid, CompIdx.Index<T>())).Returns(inst);
-                allRefTypes.Add(CompIdx.Index<T>());
+                allRefTypes.Add(reg);
             }
 
             Setup<OrderAComponent>(out var instA);
             Setup<OrderBComponent>(out var instB);
             Setup<OrderCComponent>(out var instC);
 
-            compFacMock.Setup(m => m.GetAllRefTypes()).Returns(allRefTypes.ToArray());
+            compFacMock.Setup(m => m.GetAllRegistrations()).Returns(allRefTypes.ToArray());
 
             entManMock.Setup(m => m.ComponentFactory).Returns(compFacMock.Object);
             var bus = new EntityEventBus(entManMock.Object);